feat(server): add Rustume Cloud auth, storage, and resume API
Enable optional cloud mode via RUSTUME_CLOUD with WorkOS AuthKit sessions,
Postgres migrations, authenticated resume CRUD/import, and a modular server
layout split from the former monolithic main.rs.
