+++ .osv-scanner.toml
@@ -17,3 +17,10 @@ reason = "bincode — transitive via typst -> syntect; blocked until syntect dro
 id = "RUSTSEC-2024-0436"
 ignoreUntil = 2026-10-04
 reason = "paste — transitive via typst -> hayagriva; informational (unmaintained), functionally stable"
+
+# --- SQLx: optional MySQL driver artifact in lockfile (postgres-only build) ---
+
+[[IgnoredVulns]]
+id = "RUSTSEC-2023-0071"
+ignoreUntil = 2026-10-04
+reason = "rsa — optional sqlx-mysql driver in Cargo.lock; disabled via sqlx default-features=false (postgres-only)"
