+++ crates/server/src/auth/workos.rs
@@ -1,0 +1,145 @@
+use reqwest::Client;
+use serde::Deserialize;
+
+use crate::db::User;
+
+const WORKOS_API_BASE: &str = "https://api.workos.com";
+
+#[derive(Clone)]
+pub struct WorkOsClient {
+    http: Client,
+    client_id: String,
+    api_key: String,
+}
+
+impl WorkOsClient {
+    pub fn new(client_id: String, api_key: String) -> Self {
+        Self {
+            http: Client::new(),
+            client_id,
+            api_key,
+        }
+    }
+
+    pub fn client_id(&self) -> &str {
+        &self.client_id
+    }
+
+    pub fn authorize_url(&self, redirect_uri: &str, state: &str) -> String {
+        format!(
+            "{WORKOS_API_BASE}/user_management/authorize?response_type=code&client_id={}&redirect_uri={}&provider=authkit&state={}",
+            url_encode(&self.client_id),
+            url_encode(redirect_uri),
+            url_encode(state),
+        )
+    }
+
+    pub async fn authenticate_with_code(
+        &self,
+        code: &str,
+        ip_address: Option<&str>,
+        user_agent: Option<&str>,
+    ) -> Result<WorkOsUser, WorkOsAuthError> {
+        let mut body = serde_json::json!({
+            "client_id": self.client_id,
+            "client_secret": self.api_key,
+            "grant_type": "authorization_code",
+            "code": code,
+        });
+
+        if let Some(ip) = ip_address {
+            body["ip_address"] = ip.into();
+        }
+        if let Some(ua) = user_agent {
+            body["user_agent"] = ua.into();
+        }
+
+        let response = self
+            .http
+            .post(format!("{WORKOS_API_BASE}/user_management/authenticate"))
+            .json(&body)
+            .send()
+            .await
+            .map_err(|err| WorkOsAuthError::Transport(err.to_string()))?;
+
+        if !response.status().is_success() {
+            let status = response.status();
+            let body = response.text().await.unwrap_or_default();
+            return Err(WorkOsAuthError::Api {
+                status: status.as_u16(),
+                body,
+            });
+        }
+
+        let payload: AuthenticateResponse = response
+            .json()
+            .await
+            .map_err(|err| WorkOsAuthError::Transport(err.to_string()))?;
+
+        Ok(payload.user)
+    }
+}
+
+#[derive(Debug, Deserialize)]
+struct AuthenticateResponse {
+    user: WorkOsUser,
+}
+
+#[derive(Debug, Clone, Deserialize)]
+pub struct WorkOsUser {
+    pub id: String,
+    pub email: String,
+    pub first_name: Option<String>,
+    pub last_name: Option<String>,
+}
+
+impl WorkOsUser {
+    pub fn display_name(&self) -> Option<String> {
+        match (&self.first_name, &self.last_name) {
+            (Some(first), Some(last)) => Some(format!("{first} {last}")),
+            (Some(first), None) => Some(first.clone()),
+            (None, Some(last)) => Some(last.clone()),
+            (None, None) => None,
+        }
+    }
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum WorkOsAuthError {
+    #[error("WorkOS request failed: {0}")]
+    Transport(String),
+    #[error("WorkOS API error ({status}): {body}")]
+    Api { status: u16, body: String },
+}
+
+pub async fn upsert_user(
+    pool: &sqlx::PgPool,
+    workos_user: &WorkOsUser,
+) -> Result<User, sqlx::Error> {
+    sqlx::query_as::<_, User>(
+        r#"
+        INSERT INTO users (workos_id, email, name, plan)
+        VALUES ($1, $2, $3, 'free')
+        ON CONFLICT (workos_id) DO UPDATE
+        SET email = EXCLUDED.email,
+            name = EXCLUDED.name,
+            updated_at = now()
+        RETURNING id, workos_id, email, name, plan, stripe_customer_id, created_at, updated_at
+        "#,
+    )
+    .bind(&workos_user.id)
+    .bind(&workos_user.email)
+    .bind(workos_user.display_name())
+    .fetch_one(pool)
+    .await
+}
+
+fn url_encode(value: &str) -> String {
+    value
+        .chars()
+        .map(|ch| match ch {
+            'A'..='Z' | 'a'..='z' | '0'..='9' | '-' | '_' | '.' | '~' => ch.to_string(),
+            _ => format!("%{:02X}", ch as u8),
+        })
+        .collect()
+}
