+++ crates/server/src/db/mod.rs
@@ -1,0 +1,3 @@
+pub mod models;
+
+pub use models::*;
